Don't abort when the story file is not an absolute path.
This is not a perfect fix: previously, when !include was used, a
complete path to the story had to be given or an assertion would be
triggered.  Now there is no such assertion, and the !include works if
the game is given either an absolute path or a relative path.  However,
the file to be !included will not be found unless the story file is
given either as an absolute path or in the current directory.  E.g.:

This is OK:
$ geas /path/to/file.asl

This is OK
$ geas file.asl

This is not OK
$ geas path/to/file.asl

These will all run, but the last will not find any !included files.

svn-id: r559