Fix grid window unicode input and Glk input array (un)registration.
Fixes a bug in Unicode input for grid windows that was discovered with
the inputeventtest.ulx unit test (see GitHub glk-dev project). This bug
caused fatal errors in both git and glulxe interpreters when trying to
read input from the status window and the window was less than 80
characters wide.

Caused by code registering the input array with the Glk dispatch
mechanism using a different length than the array’s original length.
The array length was modified when truncating input buffer for a grid
window as, per spec, input in grid windows is restricted to a single
line.