Don't change to the game's directory when launching.
Changing the directory causes relative paths to break:

gargoyle dir/game.z5

This would chdir into "dir", and then call the interpreter with
"dir/game.z5" as the argument, which obviously doesn't work.

I'm not sure what the rationale was behind the chdir() call, and other
Unix Glk implementations don't do this. When loading from SND/PIC files
(which I suspect nobody uses anyway), garglk makes use of gli_workdir,
so it doesn't need to be in the same directory as the game file.

Removing chdir() unless a reason is found for why it is necessary.