Increase internal linkage in garglk.
Versions of gcc prior to 10 were fine with multiple tentative
definitions, but with 10 they now become linker errors. In fact, these
variables are *not* meant to be related, but instead are meant to be
confined to their own files. It just so happens that some of their names
overlap, so they were being shared. The intent is *not* to share, so
it's correct to give them internal linkage.