Revert changes to main.c obviated by 5f7cb6a669a750652693460c7e435551fab49c7c
Since after 5f7cb6a669a750652693460c7e435551fab49c7c the working dir is
not changed, there is no longer a reason to canonicalize the game path.
The behavior this commit reverts causes weird behavior especially with
symlinks:
Files expected to be in the 'current' dir of the supplied game path
cannot be found if the the game is a symlink to a file in another
directory. This behavior is unexpected, and causes trouble in sandboxes
where all the files might be symlinks that point to files in different
directories.

Revert "Fix b7b7924 to support win32"

This reverts commit 62c6c2fdffd429217cd2b99549e6efa47c928811.

Revert "Expand game filename to full path before startup"

This reverts commit b7b79243997a6a6f8a8a0db2ebeab2f63e864351.