Be a lot more specific when searching for fonts
Previous to this, Gargoyle was not as strict in what it asked for
regarding fonts. As an example, when asking for the font:

Source Code Pro:style=Regular

On my system, the following fonts were returned:

/usr/share/fonts/truetype/sourcecodepro/SourceCodePro-Regular.otf
/usr/share/fonts/truetype/sourcecodepro/SourceCodePro-Semibold.otf
/usr/share/fonts/truetype/sourcecodepro/SourceCodePro-Black.otf
/usr/share/fonts/truetype/sourcecodepro/SourceCodePro-ExtraLight.otf
/usr/share/fonts/truetype/sourcecodepro/SourceCodePro-Light.otf
/usr/share/fonts/truetype/sourcecodepro/SourceCodePro-Medium.otf

Gargoyle would just choose the first in the list. As it happens the
first here was wanted. However, for the font:

Source Code Pro:style=Italic

The following comes back:

/usr/share/fonts/truetype/sourcecodepro/SourceCodePro-BlackIt.otf
/usr/share/fonts/truetype/sourcecodepro/SourceCodePro-LightIt.otf
/usr/share/fonts/truetype/sourcecodepro/SourceCodePro-MediumIt.otf
/usr/share/fonts/truetype/sourcecodepro/SourceCodePro-It.otf
/usr/share/fonts/truetype/sourcecodepro/SourceCodePro-ExtraLightIt.otf
/usr/share/fonts/truetype/sourcecodepro/SourceCodePro-SemiboldIt.otf

The first entry here is Black Italic, which is not what is desired, but
which does match the requested style: it is Italic, after all.

Now the code is a lot more specific in what it asks for. For example,
instead of asking just for Italic, it asks for Regular Italic (along
with related variations such as Book Italic), which avoids the wrong
weights.

This fixes #392.