Better gamma correction when rendering text
This is the initial implementation. It moves gamma correction out of the
glyph copying functions, instead putting it in gli_draw_pixel() and
gli_draw_pixel_lcd(). Those functions are where colors are interpolated
based on an alpha value (which is a color in gli_draw_pixel_lcd()), and
as a result that's the appropriate place to put gamma correction.

I've also opted to use FreeType's native LCD filtering, rather than
using the manually implemented filtering within gammacopy_lcd(). It
appears that the manual filtering was originally performed because of
Freetype only recently setting an LCD filter by default. In versions
before 2.10.3, not manually choosing a filter led to there not being any
LCD filtering. I've opted to default to choosing filter weights roughly
equivalent to what Gargoyle's manual filtering used to do, but only if
the 'gamma' setting is left at 1.0. If the user has selected a manual
gamma value for their display, I've opted to use FreeType's 'Light'
filter, which is both the sharpest and most visually correct when gamma
correction is applied.

One downside to this initial implementation is that since drawing image
graphics reuses the same pixel drawing code, there are sometimes severe
artifacts on such images when they include dark gradients. This is due
to the resolution of the gammainv[] array, which is the inverse gamma
ramp.

In the FreeType Demos source code, the equivalent array is larger, and
the equivalent to gammamap[] contains unsigned shorts instead of
unsigned chars (with values scaled to be 0 to 1023 instead of 0 to 255).

When I initially implemented this commit, I had done the same thing -
but it didn't work. To make sure I was doing the basic math and indexing
correctly, I opted to have both arrays contain 256 unsigned chars.

Since that is obviously working, my next commit will be either to
increase the size of gammainv[] (and make gammamap[] contain the
appropriately scaled shorts), or to add alternative pixel and bitmap
drawing functions which don't lead to the gamma correction code (naming
them the original function names, and renaming the new gamma-correct
functions so that they are only called by gli_draw_string() and
variants).