Remove notice about dark themes and gamma from ini
With proper gamma correct rendering of text, light text on a dark
background will automatically look thicker, and essentially be similar
in appearance to if they had previously used a gamma lower than 1.0
while using a dark theme. Now, if they were to follow the advice given
in that comment, they would actually see the opposite result from what
they were told to expect.

Also of note, is that no real display has a gamma that is below 1.0, and
very few have one of even 1.0 to begin with. However, many environments
render fonts at a gamma of 1.0 or very near there, with some Windows
programs defaulting to roughly 1.2 and others at I think 1.6 (it's been
a while since I looked at that). I've thus opted to change the list of
values to try to include:

1. The default (1.0),
2. The value for older Mac CRTs (1.8), and
3. The value typical for PC CRTs (2.2).

The reason why I specify 'CRTs' is because most LCD and other displays
try to model their response curves off of that of a CRT, even if only
roughly or loosely. Both the BT.1886 and sRGB standards are also meant
to more or less emulate the look and feel of a CRT.

I'm pretty sure most PC monitors and many HD (and some UHD) TVs try to
more or less appear similar to an old 2.2 gamma CRT, and I've heard that
modern Macs still try to roughly have a gamma of 1.8 (but have no way to
confirm that, as I have no access to Mac hardware. I've also heard the
opposite, that most Macs now use a gamma of 2.2.. So who knows what the
reality of the situation is).