Allow system libraries to be used instead of the bundled libraries
The tentative plan is to remove the bundled libraries at some point and
just use system libraries. All required libraries easily build and
install with MinGW. DLLs will still be included with the binary
distribution, but not with the source. This allows the latest versions
of libraries to be used with limited pain.