Update the AppImage builder
Ubuntu 16.04 includes an older gtk3 which does not have the new
functions used here. Switch to 20.04.

It also appears that ftjam is not in the i386 repositories for 20.04, so
I've removed 32-bit AppImage support for the time being. Once the switch
to CMake is (eventually) done, 32-bit support can probably come back.