Fix fonts being subpixel filtered twice
Gargoyle does its own subpixel filtering of rendered fonts. This used to
work fine with versions of FreeType prior to 2.10.3, but 2.10.3 and up
use a default set of filter weights.

Previously, one would need to explicitly set the filter weights for any
filtering to occur, leaving Gargoyle free to perform the filtering
itself.