Update ini files for new options and gamma changes
The gamma correction documentation no longer needs to recommend values
below 1.0 for light text on dark backgrounds, so has been adjusted.

Additionally, documentation for the options added in pull request #433
has been added, with the 'lcd' option being moved to a new section of
the default config files (which is where the documentation resides).