Use more precise math when blending color values
Change summary:

1. Changed 'GAMMA_SHIFT' to 'GAMMA_BITS' to more clearly portray its meaning
2. Changed 'gammamax' to 'GAMMA_MAX', #defined for use in other constants
3. Changed 'mul255()' and 'mulhigh()' to use more precise math

Rationale for changes:

In the past, the 'mul255()' macro has been used to perform a 'multiply'
operation on two pixel values, such that the two values multiply as if they were
in the range of 0.0 to 1.0. This has basically only been used as part of the
math for blending two colors based on an alpha value.

When I implemented gamma correction, I added an additional 'mulhigh()' macro
that behaved in much the same way, but with a maximum value that was determined
by the 'GAMMA_SHIFT' defined constant. 'GAMMA_SHIFT' was made because that's how
it's done in the FreeType demos' source code, but what it represents is how many
more bits linear light values can use compared to standard 8-bit light values.

I wrote the 'mulhigh()' macro to use the same basic formula as the 'mul255()'
macro, mostly because I was lazy and didn't want to double-check to see if there
was any better way to do the math. The math for 'mul255()' did work, after all.

However, after some experimentation, I've found that there's a more accurate way
to implement both macros. I've also put some additional thought into using the
term 'GAMMA_SHIFT', as it slightly obscures what the value is actually meant to
represent: the number of bits to use for linear light values.

Math Overview

The old math for the 'mul*()' macros was as follows:
(a * (b + 1)) / 256

However, the division at the end was implemented as '>> 8', which likely is a
little bit faster than division. However, dividing by 256, when the values are
actually in the range of 0 to 255, is what makes the 'b + 1' part necessary.

While marginally slower since it lacks bitshifting, this is much more accurate:
(a * b + 127) / 255

Note that while the original method used '(b + 1)', this new method does not use
parentheses around the 'b' and the value being added. Instead, the value (which
is now 127) is being added After 'a' and 'b' are multiplied together.

This is because to be honest, the '+ 127' isn't actually necessary for accuracy,
but without it (a * b / 255) the results would essentially be 'floored'; that is
to say, the following two are equivalent:

1. a * b / 255
2. floor(((float)a / 255.0) * ((float)b / 255.0) * 255.0)

By adding 127 to the product of 'a' and 'b', however, the result will be rounded
instead of floored. It's essentially the same as adding 0.5 before flooring.