Support building TADS with LLP64
Specifically TADS assumes that a long is large enough to hold a pointer,
but on LLP64, it's not. This causes a build error with
x86_64-w64-mingw32.

This probably should be uintptr_t but that makes format strings ugly and
unsigned long long is fine until 128-bit systems come out.