Use the STL instead of custom data structures
This necessitates a switch to C++ for this file. At the present moment
it's still a mix of C- and C++-style code, with the latter being used so
that the STL's map can be used instead of custom data structures.

Rationale: Custom data structures are more error-prone than the STL, and
make the code harder to read. In addition, a bundled hash implementation
(that could fall out of date easily) is no longer needed. The less code
Gargoyle itself has, the better.