Split MinGW toolchains into two separate files
This is suboptimal, but apparently CMake does not cache variables when
running toolchains, so the default (i686) is used in later runs no
matter what. As a result, things happen such as picking up the wrong
pkg-config files. Ultimately, if the 64-bit versions are installed but
the 32-bit (or system) ones found, it "works", because the library
search path is set by other means, but that's just incidental: if a
library is not installed, but the wrong one is found, the build fails.

Now the user is required to specify the toolchain file for the target
architecture.

See:

https://stackoverflow.com/questions/28613394/check-cmake-cache-variable-in-toolchain-file

and

https://gitlab.kitware.com/cmake/cmake/-/issues/16170