Use the STL for event handling data structures
As with previous STL changes, this removes a lot of custom data
structures in order to greatly simplify the code.

The behavior here is slightly changed: in the earlier version,
glk_select() would prioritize all external (input) events before all
internal events. That is, if a window resize event arrived and then an
input event, glk_select() would return the input event first and then
the resize event.

Now events are dispatched in the order received. This won't matter to
properly-written Glk programs, because they cannot assume what kind of
event will be returned by glk_select(), per §4 of the Glk spec: “This is
why you must always call glk_select() in a loop, and continue the loop
until you get the event you really want.”