Use Qt to determine the path of the executable
QCoreApplication::applicationDirPath() returns the path to the
executable, so get rid of a lot of platform-specific code here. One
caveat: for systems other than Linux/Windows/Mac, this uses argv[0] to
find the executable, which means if execve() is called with an
"incorrect" first argument, the path won't be found. That doesn't matter
since nobody will do that, and if they do, I'm not sure it's worth
catering to.