Only set _XOPEN_SOURCE
The previous comment to this implied that MinGW required _POSIX_C_SOURCE
to expose POSIX functionality, and that _XOPEN_SOURCE wasn't supported.
For the latest MinGW, at least, this isn't true. It exposes its
POSIX-ish interface regardless of whether the feature test macros are
used (even if -std=c11 or similar is passed). The only thing
_POSIX_C_SOURCE is used for is determining whether to use the so-called
ANSI stdio, but this is also activated by _XOPEN_SOURCE.

In short, if this was ever right, it's not anymore. Just use
_XOPEN_SOURCE to select the XSI extension for SUSv3, which appears to be
sufficient on all platforms.