Copy TTF files to the CMake build directory
This is a direct consequence of the removal of bundled fonts. Without
them, Gargoyle has to find fonts somewhere, and while that's not
normally a problem, for users who are just experimenting with Gargoyle,
the fonts won't exist if it's not installed. By copying these fonts, at
least running directly from the build directory (e.g. "cd build &&
./terps/bocfel zork1.z3") should work.

It might be useful to have an alternative approach to finding fonts: for
now, if the user-specified font doesn't exist, and the fallback TTF
files can't be found, then Gargoyle will just fail to start (with an
error message pointing to the fonts). It might be more user-friendly to
find a font, any font, when the "proper" fonts don't exist. Fontconfig
allows aliases such as "monospace" and "serif" to be used, and there are
likely similar mechanisms on Windows and Mac.