Centralize config file discovery
The config file is needed in both the launcher and the Glk library
itself. The launcher uses the config file to possibly determine which
interpreter to use, while the Glk library uses it for its configuration
in general. That is to say, the "terp" entries in the config file are
used by the launcher, and the rest by garglk.

Gargoyle searches in several places for its config file, and both the
launcher and garglk had code to handle lookups. However, they disagreed
slightly on how to look things up.

The new code adds a function which returns all possible config file
locations so both the launcher and garglk will always get the same list
of config files.