Ensure only the relevant modifiers are present
Specifically, don't check for modifiers using bitwise "and": this checks
whether one modifier is set, but doesn't care about the status of other
modifiers. As a result, any number of modifier combinations could result
in the same event; for example, Control-A, Control-Shift-A, and
Control-Alt-A would all match with code like:

    if (modifiers & ControlModifier)

The obvious solution is something like:

    if (modifiers == ControlModifier)

but it's not always that simple. There are "virtual" modifiers, such as
NSFunctionKeyMask on macOS, which is set when "function" keys are
pressed, which includes F-keys as well as others, like Home and End.  So
it's not enough to check equality, either.

The solution is to convert the modifiers to just the "interesting"
modifiers. In this case that's shift, alt/option, command, and control.
Once masked, equality tests can be used with no problem.

Qt is similar. It includes Qt::KeypadModifier if a keypad key is
pressed, so it's masked to shift, control, alt, and meta.