Rename VERSION
This is the result of a bug on macOS, where its cstddef header does:

    #include <version>

This picks up the VERSION file in the cheapglk directory, which is
obviously the wrong file. There's no reason at all a system header
should be looking in a project's directory for its own headers, but
there you are.