Undefine bzero before defining it
Implementations should not be exposing bzero() unless strings.h is
included, which it isn't, but at least Haiku unconditionally pulls in
strings.h if you include string.h. Since Haiku provides bzero() as a
macro instead of a function, the macro here clashes with it. Undefine
first to prevent problems.