Check file contents to determine which interpreter to use
Not all formats can be reliably probed like this, but for those which
can, prefer the probe over extensions. Extensions are still used as a
fallback, and if the user has specified an interpreter for a particular
file/file type in garglkrc, that will be preferred over anything else.

This should be more robust for file managers/operating systems which
rely more on file content as opposed to extensions to determine file
type.