Remove overrides which serve no functional purpose
The Trinity override is completely necessary since it fails to start
with fewer columns.

The Winter Wonderland override is necessary to make the ASCII art come
out right. More subjective, but probably OK to have.

Git & Glulxe at 80x25 is arbitrary. Why just Glulx games?

Hugo: The border under the status bar is different than other
interpreters; it's more jarring to have it in Hugo than not.

Magnetic: These games work fine with the defaults.

Agility: This was commented out already anyway.

Curses: The game works fine with the defaults.

Photopia: The game works fine with the defaults.

It's actually possible that some parts of the listed games do fail at
the defaults; I tested the first parts to see if there were problems as
in Trinity. But I think if there are problems with the default 60x25, it
should be changed rather than trying to override it. 80x25 is a
reasonable default, given that so many games were written with that size
in mind, but for now, that'd be a pretty big change. This will have some
effect on existing users (assuming they have the default garglk.ini
installed), but I think it's best to not make subjective choices about
specific games.