Allow native file dialogs to be disabled with Qt
The primary reason for this is to support Flatpak. With native dialogs
(at least with Plasma), files aren't properly mapped to the Flatpak
sandbox. They can be selected, but the filename returned by QFileDialog
is a non-existent file. Using non-native file dialogs doesn't have this
problem.

This directly contradicts the advice at
https://docs.flatpak.org/en/latest/portals-qt.html. I don't know why.