Reintroduce IDI_GDOC
This icon isn't used by name, but is used by number in the registry in
order to provide a document type icon for the various story files.
Documentation on this isn't exactly good, but it seems that the icons
are ordered by name, so now that IDI_GDOC and IDI_ICON1 are the only
ones, IDI_GDOC has the number 0, and IDI_ICON1 the number 1. Qt uses
IDI_ICON1 by name, so that's OK, but the installer uses the index, so
make sure it's in line with what actually exists.