Don't set _XOPEN_SOURCE when building C++
This goes down a rabbit hole of POSIX, C++, and individual
implementations.

For historical reasons, standard C headers expose POSIX functions that
are not also standard C functions: stdlib.h on POSIX, for example,
includes a function called drand48. This is not a standard C function.
So compilers, when told to build as C99 (e.g. -std=c99 with gcc), set
macros which instruct the C library to not expose this function. On
glibc, this looks something like:

    #if defined __USE_XOPEN
    extern double drand48 (void);
    #endif

__USE_XOPEN is _not_ set with -std=c99 unless requested via feature test
macro such as _XOPEN_SOURCE.

All well and good: a standard C (not POSIX) program can use -std=c99 and
non-standard names won't be pulled in; then, if POSIX is wanted,
_XOPEN_SOURCE can be set to expose POSIX functionality.

So a program wanting to conform to C99 and SUSv3 does this:

    gcc -std=c99 -D_XOPEN_SOURCE=600

And now, C++. On at least Linux/glibc, FreeBSD, and OpenBSD, the C
library headers do _not_ provide the same courtesy to C++ compilers. If
g++ is given -std=c++14, the C libraries _still_ expose non-standard
functions. The tests which check for strict C compiliance do not care
about strict C++ compliance. As a result, even with -std=c++14, the
implementation might expose names which are not reserved, potentially
clashing with the user's program.

On to feature test macros with C++. As it stands, these are unnecessary
because, as noted above, implementations apparently expose POSIX
functionality (plus more!) even without feature test macros being set.
Moreover, setting feature test macros can even be harmful. According to
Oracle Solaris documentation [1]:

“C++ bindings are not defined for POSIX or SUSv4, so specifying feature
test macros such as _POSIX_SOURCE, _POSIX_C_SOURCE, and _XOPEN_SOURCE
can result in compilation errors due to conflicting requirements of
standard C++ and these specifications.”

This also manifests itself on FreeBSD and OpenBSD.

On FreeBSD, if <iostream> is included:

    $ clang++ -std=c++14 -D_POSIX_C_SOURCE=200112L t.cpp
    In file included from t.cpp:1:
    In file included from /usr/include/c++/v1/iostream:37:
    In file included from /usr/include/c++/v1/ios:215:
    /usr/include/c++/v1/__locale:631:16: error: use of undeclared identifier 'isascii'
            return isascii(__c) ? (__tab_[static_cast<int>(__c)] & __m) !=0 : false;

Or on OpenBSD, if <fstream> is included:

    $ clang++ -std=c++14 -D_XOPEN_SOURCE=600 t.cpp
    In file included from t.cpp:1:
    In file included from /usr/include/c++/v1/fstream:183:
    In file included from /usr/include/c++/v1/ostream:138:
    In file included from /usr/include/c++/v1/ios:215:
    In file included from /usr/include/c++/v1/__locale:32:
    In file included from /usr/include/c++/v1/support/newlib/xlocale.h:25:
    /usr/include/c++/v1/support/xlocale/__strtonum_fallback.h:23:64: error: unknown type name 'locale_t'
                                                    char **endptr, locale_t) {

So there's apparently no proper way to get strict C++ plus POSIX. The
best we can do is just hope that implementations don't provide
non-standard names that clash with anything Gargoyle uses. So far that
appears to be the case, at least.

[1]: https://docs.oracle.com/cd/E88353_01/html/E37853/c-plus--plus--7.html