SCOTT: Implement fixes from asdfcurs This a fork of ScottFree available at https://www.ifarchive.org/indexes/if-archive/scott-adams/interpreters/scottfree/
Quote from README:

"If action 52 (Get, with a check on carrying capacity) fails--i.e. you
don't get it because you've got too much--now abort the rest of the
action line.  This is the behavior of the VIC-20 driver, and it makes
sense in every case I can see within games I have in ScottFree format
(with the possible exception of MAKE FUEL in Buckaroo Banzai).

If you break you neck by moving an invalid direction in the dark, action
61 (kill player) is executed, instead of just ending the game.  You may
have a chance to live again once killed, as in Adventureland.

Fixed a bug that would only stop you from getting items if you were
exactly at your carring capacity.  Situations can arise when you are
over capacity (such as chigger bites in Adventureland)."