Fix usages of glkunix_fileref_get_filename
Unlike the earlier Gargoyle-specific function, this one is defined in
glkstart.h, a Unix Glk-specific header (which includes garglk). As such,
this header must be included to get the GLKUNIX_FILEREF_GET_FILENAME
macro.

One small problem: both Alan2 and TADS want to use this functionality,
but also include their own copy of glkstart.h. There's no reason for
them to have this, since the Glk implementation ought to be the one
providing it. As such, these versions have been removed from the
repository.

This also means that a test against GARGLK is needed before including
glkstart.h. For portability this should be more than just GARGLK, but in
this repository, it's fine. For use outside of Gargoyle, this should be
more specific, e.g. testing a macro for whether it's Unix Glk, Windows
Glk, etc. Right now, GARGLK is a proxy meaning "Unix Glk", used because
these interpreters don't care about anything but Gargoyle.