Sanitize config entries
Before this, there was no sanitization of config values, so the user
could set things like -50 columns, or a line spacing of -100. A lot of
these invalid values could lead to segfaults, while others caused
obvious graphics glitches.

All (hopefully) config entries are now checked to ensure they're at
least positive, where necessary. In addition, parsing of colors is more
correct: earlier, a value such as "-b-b-b" was allowed, and is not any
more.

It's possible that this causes behavior change, but only in config files
that were clearly wrong, so that's fine.