SCOTT: Add special handling for Mysterious Adventures
This detects the "extracted" dat versions of Brian Howarth's 11 Mysterious
Adventures and switches on the SCOTTLIGHT and PREHISTORIC_LAMP options
for them, for correct lamp behaviour.

SCOTTLIGHT gives more precise warnings every five turns when the lamp is
nearing depletion ("Light runs out in 5 turns.")

PREHISTORIC_LAMP will remove the lamp from play when it runs out.
If this is not set, as is the case when running these games with the
ScottFree version currently in master, you end up with two lamps in
your inventory, one lit and one unlit.

This is not properly tested yet but seems to work.