Use appropriate types
The key here is a pair of characters, which are generally represented as
an unsigned type (FT_ULong for FreeType, glui32 for Gargoyle); but prior
to this, the keys were being stored ints. Generally speaking this
shouldn't matter: int is almost surely 32 bits which, even signed, will
fit all Unicode characters. But semantically it's wrong (and technically
it's wrong since INT_MAX can be as low as 32767, though I suspect there
will be many, many other problems if that's the case).