Don't artifically reverse User1 and User2 styles
Originally these were set to be opposite colors from the rest of the
styles, presumably for interpreters to use as status bars. However, the
garglk.ini that's shipped doesn't reverse the colors, and it looks like
all interpreters that use user styles set the reverse video bit,
negating the "advantage" of having artificial reverse video.

This brings Gargoyle's code in line with the config file, and should
make interpreters look better in the absence of a config file.