Avoid unnecessary calculation
gli_draw_pixel used to be used in various places, but it's been
replaced by more specialized functions. As a result, the single caller
to it now always passes 0xff as the alpha channel, resulting in a
calculation that is always 0, so remove the alpha channel parameter and
the calculation.