Add optional KDE support for better text editing
This optional feature allows the user's preferred text editor to be
discovered and used to edit the config file. Qt's support assumes that
the operating system will be able to open whatever config file is thrown
at it (typically .garglkrc or garglk.ini), but that's not necessarily
going to be correct. KDE allows the user's preferred handler for
text/plain to be directly discovered and run, regardless of what the
desktop thinks it ought to use to open the file. This mimics the Mac
behavior (which uses "open -t").