Don't set _XOPEN_SOURCE
Now that everything (more or less) is C++, _XOPEN_SOURCE is pretty
pointless, since it was set for C files anyway. There are a few C files
in cheapglk still, but those don't expect POSIX, so there's no point in
this.

The launcher is all C++ and I botched the change way back when anyway,
setting _XOPEN_SOURCE for garglk twice, instead of garglk and gargoyle.