Allow quoted filenames in overrides
For game-specific overrides, the syntax is something like this:

    [game1.z5 game2.z5 game3.z5]
    config value

This way you can set overrides for many games at once. However, it also
means that you can't match a filename with spaces in it, as each word
would be treated as a separate game.

Now, files can be quoted so that they can contain spaces.