PLUS: Minor fixes from upstream
Remove incorrect cast
MemAlloc() takes a size_t in Plus, but an int in ScottFree, which is inconsistent and causes
things like this.

Fix a few small type mismatches
uint64_t isn't guaranteed to be an unsigned long long, so cast in printf
and use SCN macro in scanf.

int16_t isn't guaranteed to be a short, so use an SCN macro in scanf.

enums aren't guaranteed to have the same representation as int, so use
an intermediate variable.