Use position-independent code for garglkmain if it's used for garglk
garglkmain is a sort of glue code which provides the "main" function for
interpreters (whose entry point is glk_main). It's always built as a
static library, even if libgarglk is build shared. As such, it won't be
built as PIC, even if libgarglk is. That's been fine for a while, but
broke with Clang 15: building Gargoyle with gcc, then linking it to a
project with clang caused relocation errors. But building _everything_
PIC caused problems on Haiku. Right now, the solution is to build
garglkmain the same as garglk, since garglkmain needs to be explicitly
linked in by third-party interpreters: the launcher uses libgarglk but
is _not_ an interpreter, and therefore provides its own "main" function.

This doesn't affect Gargoyle itself, since it won't be mixing
compilers/linkers. It only affects anybody building against Gargoyle
with their own interpreter, linking to Gargoyle's libraries.

The "real" solution would probably be to have garglkmain be shared as
well, and at the same time recommend that official builds be static
(there's not really much gain to having libgarglk be shared: a bit of
disk space is saved but you rarely run multiple interpreters
simultaneously, rendering memory savings moot, and it's a pretty small
library anyway).

But this is the path of least resistence for the time being.