Avoid passing around exedir
exedir was/is used to find garglk.ini on Windows. Due to historical
reasons, garglk.ini from the install directory (e.g.
C:\Program Files\Gargoyle) exists as a user config file. Modern Gargoyle
releases use a more appropriate location, but this is still loaded for
backward compatibility.

Till now, all platforms were calculating exedir, the location of the
Gargoyle main binary, and passing it around, though it is only _used_
when building for Windows.

This change removes that passing around, instead calling for the exedir
only where it's required. This pushes the "find application dir" logic
to the sys* files. It's implemented on Mac even though it doesn't need
to be (since it won't ever be called there), just for completeness.

At the same time, rename the entirely-without-information variable
"path" to "interpreter_dir". There was a conflation of the interpreter
dir and the binary dir, because all platforms need the interpreter dir,
to actually find the interpreters, and Windows was kind of abusing this
by "knowing" that this was also the executable path, as well. On Unix
platforms, for example, these two are generally different (e.g. /usr/bin
and /usr/libexec/gargoyle), so exedir was actually getting passed the
interpreter dir, but in the end it's not _used_ on Unix, so it didn't
matter. But it's confusing, so hopefully this makes things a bit more
clear.