Have default themes (dark/light) be files, not hard-coded C++ structures.
This has a big advantage: modifying the system themes (in the source
code) is a lot easier. These are still built in to the Gargoyle binary,
so that Gargoyle will continue to work regardless of whether the theme
files exist, but they are built in based on the JSON files. They're also
installed, now, so that users have a template to work from if they just
want to tweak the dark/light themes.

One disadvantage: invalid dark/light themes will not be caught at
compile-time, as they have to be parsed as JSON. This is probably
fine, given that we control the contents of these files.