Convert to unsigned char before storing
char can be (and usually is) signed, but dwin->chars is unsigned. As a
result, Latin-1 input that's above 127 (i.e. no longer ASCII) is
represented by a *negative* number. This is fine when chars are all
that's involved, but since a char is being assigned to a glui32, the
negative value is converted to a very large 32-bit value. Cast to
unsigned char to get a value in the range [0, 255], properly
corresponding to Latin-1.